% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuanShiToolsRfunctions.R
\name{fun1}
\alias{fun1}
\title{quiz 2 function 1}
\usage{
fun1(a, x)
}
\arguments{
\item{a}{matrix}

\item{x}{vector}
}
\value{
object
}
\description{
Write an R function that, given a numeric matrix A and a numeric vector x, calculates xTA−1x
}
