% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuanShiToolsRfunctions.R
\name{fun3}
\alias{fun3}
\title{quiz 2 function 3}
\usage{
fun3(a)
}
\arguments{
\item{a}{matrix}
}
\value{
matrix
}
\description{
Write a function that takes a numeric matrix and standardizes its columns as (x−mean(x))/sd(x).
}
