% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuanShiToolsRfunctions.R
\name{\%fun2\%}
\alias{\%fun2\%}
\title{quiz 2 function 2}
\usage{
a \%fun2\% x
}
\arguments{
\item{a}{matrix}

\item{x}{vector}
}
\value{
object
}
\description{
rewrite function for the preceding problem so it is a binary operator rather than an apparent function call, that is, if your function from the preceding problem was invoked
}
