% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuanShiToolsRfunctions.R
\name{myapply}
\alias{myapply}
\title{Homework 2 q5 function}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{MARGIN}{margin of object you want to work on}

\item{FUN}{function}

\item{x}{matrix}
}
\description{
question 5
}
