% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuanShiToolsRfunctions.R
\name{selectmydata}
\alias{selectmydata}
\title{wrapper for acquiring marginals and other slices of data data2007}
\usage{
selectmydata(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
list
}
\description{
select two speific rows and compute the product
}
